requirements.txt
kafka-python
pyspark
pyyaml
python-snappy
